temp

+++ db/migrate/20170807095855_create_chapters.rb
@@ -3,7 +3,7 @@ class CreateChapters < ActiveRecord::Migration[5.1]
     create_table :chapters do |t|
       t.belongs_to :subject, index: true
       t.string :name
-      t.integer :level
+      t.float :level
       t.boolean :onExam
       t.boolean :onStudy
